src/main/java/ru/zrs/easy/ValidParentheses.java
package ru.zrs.easy;

import java.util.Stack;

public class ValidParentheses {

    public static boolean checkValidParentheses(String str) {

        Stack<Character> stack = new Stack<>();
        for (char c : str.toCharArray()) {
            if (c == '(')
                stack.push(')');
            else if (c == '{')
                stack.push('}');
            else if (c == '[')
                stack.push(']');
            else if (stack.isEmpty() || stack.pop() != c)
                return false;
        }
        return stack.isEmpty();
    }
}
